@@ -36,26 +36,37 @@
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.23.0",
     "body-parser": "^1.17.0",
+    "connect-redis": "^3.2.0",
+    "cookie-parser": "^1.4.3",
     "css-loader": "^0.19.0",
     "css-modules-require-hook": "^4.0.5",
     "dotenv": "^4.0.0",
     "eslint": "^3.16.1",
     "eslint-plugin-react": "^6.10.0",
     "express": "^4.14.0",
+    "express-logger": "0.0.3",
+    "express-session": "^1.15.1",
     "fetch-api": "0.0.2",
     "html-webpack-plugin": "^2.28.0",
+    "material-ui": "^0.17.0",
+    "method-override": "^2.3.7",
     "moment": "^2.17.1",
     "node-fetch": "^1.6.3",
+    "passport": "^0.3.2",
+    "passport-google-oauth": "^1.0.0",
     "postcss-loader": "^0.6.0",
-    "react": "^15.2.0",
+    "react": "^15.4.2",
     "react-big-calendar": "^0.12.3",
     "react-bootstrap": "^0.30.7",
     "react-bootstrap-date-picker": "^4.0.0",
     "react-datepicker": "^0.41.1",
-    "react-dom": "^15.2.0",
+    "react-dom": "^15.4.2",
+    "react-google-login-component": "^0.5.2",
     "react-hot-loader": "^1.3.0",
     "react-router": "^3.0.0",
     "react-s-alert": "^1.2.2",
+    "react-tap-event-plugin": "^2.0.1",
+    "redis": "^2.7.1",
     "style-loader": "^0.12.4",
     "underscore": "^1.8.3",
     "webpack": "^1.12.2",
